# swagger.json
{"openapi":"3.0.0","paths":{"/signup":{"post":{"operationId":"AuthController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"409":{"description":"E-mail in use","content":{"application/json":{"schema":{"type":"object","properties":{"statusCode":{"type":"number","example":409},"message":{"type":"string","example":"e-mail já cadastrado em nosso sistema"}}}}}}}}},"/signin":{"post":{"operationId":"AuthController_signin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthUserDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string","example":"Bearer ..."}}}}}},"401":{"description":"Invalid Credentials","content":{"application/json":{"schema":{"type":"object","properties":{"statusCode":{"type":"number","example":401},"message":{"type":"string","example":"credenciais inválidas"}}}}}},"404":{"description":"E-mail not found","content":{"application/json":{"schema":{"type":"object","properties":{"statusCode":{"type":"number","example":404},"message":{"type":"string","example":"e-mail não encontrado"}}}}}}}}},"/update-credentials":{"patch":{"operationId":"AuthController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}}}},"/recover-password":{"post":{"operationId":"AuthController_recoverPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecoverPasswordDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"o link de recuperação foi enviado ao seu e-mail"}}}}}},"404":{"description":"E-mail not found","content":{"application/json":{"schema":{"type":"object","properties":{"statusCode":{"type":"number","example":404},"message":{"type":"string","example":"e-mail não encontrado"}}}}}}}}},"/recover-password/callback/{token}":{"patch":{"operationId":"AuthController_recoverPasswordCallback","parameters":[{"name":"token","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecoverCallbackDto"}}}},"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{"message":{"type":"string","example":"sucesso"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}}}},"/active-user/{id}":{"patch":{"operationId":"AuthController_activeUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActiveUserDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string","example":"Bearer ..."}}}}}},"400":{"description":"User already active or invalid code","content":{"application/json":{"schema":{"type":"object","properties":{"statusCode":{"type":"number","example":400},"message":{"type":"string","example":"usuário já está ativo | código inválido"}}}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"type":"object","properties":{"statusCode":{"type":"number","example":404},"message":{"type":"string","example":"usuário não encontrado"}}}}}}}}},"/google":{"get":{"operationId":"AuthController_googleAuth","parameters":[],"responses":{"200":{"description":""}}}},"/google/callback":{"get":{"operationId":"AuthController_googleAuthRedirect","parameters":[],"responses":{"200":{"description":""}}}},"/facebook":{"get":{"operationId":"AuthController_facebookAuth","parameters":[],"responses":{"200":{"description":""}}}},"/facebook/callback":{"get":{"operationId":"AuthController_facebookAuthRedirect","parameters":[],"responses":{"200":{"description":""}}}},"/resend-code/{email}":{"post":{"operationId":"AuthController_resendCodeEmail","parameters":[],"responses":{"201":{"description":""}}}}},"info":{"title":"Auth Service","description":"The auth service only...","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"username":{"type":"string","example":"any_username"},"email":{"type":"string","example":"any_email@mail.com"},"password":{"type":"string","example":"any_password"}},"required":["username","email","password"]},"UserEntity":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string","description":"the name of user","example":"any_username"},"email":{"type":"string","description":"the name of user","example":"any_email@mail.com"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}},"required":["id","username","email","created_at","updated_at"]},"AuthUserDto":{"type":"object","properties":{"email":{"type":"string","example":"any_email@mail.com"},"password":{"type":"string","example":"any_password"}},"required":["email","password"]},"UpdateUserDto":{"type":"object","properties":{"username":{"type":"string","example":"any_username"},"password":{"type":"string","example":"any_password"}},"required":["username","password"]},"RecoverPasswordDto":{"type":"object","properties":{"email":{"type":"string","example":"any_email@mail.com"}},"required":["email"]},"RecoverCallbackDto":{"type":"object","properties":{"password":{"type":"string","example":"any_password"}},"required":["password"]},"ActiveUserDto":{"type":"object","properties":{"activationCode":{"type":"string","description":"The code received in e-mail","minLength":4,"maxLength":4,"example":"0000"}},"required":["activationCode"]}}}}